requirements.txt
# Home Health AI API Requirements
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.0.0
celery>=5.3.0
redis>=4.5.0
psycopg2-binary>=2.9.0
django-extensions>=3.2.0
python-multipart>=0.0.6
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
python-decouple>=3.8
